Add abstract StructureWorkflow
Created a structure workflow class that can be mixed into workflows that might want to clean a structure in their setup (e.g. VaspWorkflow, QuantumEspressoWorkflow)
