+++ src/simmate/engine/__init__.py
@@ -25,5 +25,7 @@ from simmate.database import connect
 
 from .workflow import Workflow, workflow
 from .s3_workflow import S3Workflow
+from .staged_workflow import StagedWorkflow
+from .structure_workflow import StructureWorkflow
 from .web_api_workflow import WebApiWorkflow
 from .execution import SimmateWorker as Worker
